new adaptive harris

+++ core/vsl/vsl_canny_nms.cxx
@@ -12,7 +12,7 @@ int vsl_canny_nms(int _xsize, int _ysize,
 		  float * const * _dx, float * const * _dy, float const * const * _grad,
 		  float * const *_thick, float * const * _theta) 
 { 
-  const float k = 180.0/3.1415926;
+  const float k = float(180.0/3.1415926);
   int n_edgels_NMS = 0; // return value for this function
 
   for (int y=_ysize-2; y>0; --y) {
